backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary
python-dotenv==1.0.0